Define an API to create a character
We've had to make the `data`, `level` and `class_` fields optional in the `Character` schema, as they will be populated in the character sheet.

For now, we subit `party_id` as well, even though this isn't something we've implemented in the frontend, so we might have to set it to `1` for a while.

Fixes #49